# subscription-aggregator/docs/swagger.json
{"schemes":["http"],"swagger":"2.0","info":{"description":"REST API for managing and aggregating user subscriptions.","title":"Subscription Aggregator API","contact":{},"version":"1.0"},"host":"localhost:8080","basePath":"/","paths":{"/subscriptions":{"get":{"description":"Get all subscriptions for a user","produces":["application/json"],"tags":["subscriptions"],"summary":"List subscriptions","parameters":[{"type":"string","description":"User ID (UUID)","name":"user_id","in":"query","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/model.Subscription"}}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"post":{"description":"Create a new subscription record","consumes":["application/json"],"produces":["application/json"],"tags":["subscriptions"],"summary":"Create subscription","parameters":[{"description":"Subscription data","name":"subscription","in":"body","required":true,"schema":{"$ref":"#/definitions/model.Subscription"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/model.Subscription"}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/subscriptions/total-cost":{"get":{"description":"Calculate total cost of active subscriptions for a user in a period","produces":["application/json"],"tags":["subscriptions"],"summary":"Get total subscription cost","parameters":[{"type":"string","description":"Filter by user ID","name":"user_id","in":"query"},{"type":"string","description":"Filter by service name","name":"service_name","in":"query"},{"type":"string","description":"Start of period (MM-YYYY)","name":"from","in":"query","required":true},{"type":"string","description":"End of period (MM-YYYY)","name":"to","in":"query","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"total":{"type":"integer"}}}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"/subscriptions/{id}":{"get":{"description":"Get a single subscription by its UUID","produces":["application/json"],"tags":["subscriptions"],"summary":"Get subscription by ID","parameters":[{"type":"string","description":"Subscription ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Subscription"}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Not Found","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"put":{"description":"Update an existing subscription by ID","consumes":["application/json"],"produces":["application/json"],"tags":["subscriptions"],"summary":"Update subscription","parameters":[{"type":"string","description":"Subscription ID","name":"id","in":"path","required":true},{"description":"Updated subscription data","name":"subscription","in":"body","required":true,"schema":{"$ref":"#/definitions/model.Subscription"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Subscription"}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Not Found","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"delete":{"description":"Delete a subscription by ID","tags":["subscriptions"],"summary":"Delete subscription","parameters":[{"type":"string","description":"Subscription ID","name":"id","in":"path","required":true}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":{"type":"string"}}},"404":{"description":"Not Found","schema":{"type":"object","additionalProperties":{"type":"string"}}},"500":{"description":"Internal Server Error","schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}},"definitions":{"model.Subscription":{"type":"object","properties":{"end_date":{"description":"EndDate — опциональная дата окончания подписки в формате \"MM-YYYY\"\nЕсли не указана — подписка считается активной.","type":"string"},"id":{"description":"ID — уникальный идентификатор подписки (UUID, генерируется в БД)","type":"string"},"price":{"description":"Price — стоимость подписки в рублях (целое число, копейки не учитываются)","type":"integer"},"service_name":{"description":"ServiceName — название сервиса (например, \"Yandex Plus\")","type":"string"},"start_date":{"description":"StartDate — дата начала подписки в формате \"MM-YYYY\"","type":"string"},"user_id":{"description":"UserID — идентификатор пользователя (UUID в строковом виде)","type":"string"}}}}}